press d 7 times to show debug menu
